+++ tkintertools/__init__.py
@@ -48,7 +48,7 @@ from .__main__ import *
 from .constants import *
 
 __author__ = 'Xiaokang2022<2951256653@qq.com>'
-__version__ = '2.6.9'
+__version__ = '2.6.10.dev0'
 __all__ = [
     # Container Widgets
     'Tk', 'Toplevel', 'Canvas',